Remove routes_mapper fixture and use configurator routes
Some of our tests, mostly views tests, need the pyramid application to
be configured with a routes mapper so that `request.route_path` and
`request.route_url` work as expected.

We had a clunky fixture, `routes_mapper`, which provided a dummy route
mapper, but it's actually cleaner and easier to use the proper Pyramid
routes mapper, which is set up by `pyramid.testing.setUp`. All that is
required is that tests explicitly defined the routes they expect to
exist using `config.add_route`.

This commit removes `routes_mapper` and also cleans up the fixture
definitions so that we're not calling `pyramid.testing.setUp` for every
test unconditionally. In my testing this knocks 5-6 seconds (25-30%) off
test suite runtime.