Publish annotation model object in AnnotationEvent
Publish the actual annotation model object
(h.api.models.elastic.Annotation or h.api.models.Annotation, depending
on the postgres feature flag) with the AnnotationEvent when an
annotation is created or deleted as we used to, instead of publishing a
dict.

h/subscribers.py would now crash when the postgres feature flag is on
because it tries to JSON serialize the annotation model object, and the
Postgres model objects aren't JSON serializable. So have it use
AnnotationJSONPresenter on the object, instead of views doing it.