Index newly created annotations
Synchronously index newly created annotations into the new Elasticsearch
index, when the 'postgres' feature flag is on.

Annotation updates and deletes aren't indexed yet.

- Add a new create_annotation() method to h.api.search.Client.

  Previously this client class was only used for searching the search
  index, and legacy Elasticsearch "model" classes were used for writing
  to the index.

  The legacy classes are still used for writing to the legacy index but
  this Client class is now used for writing to the new index as well as
  searching both the legacy and new indexes.

- Add a new AnnotationEvent subscriber that receives annotation create
  events and calls the above new search client method to index the
  annotation.

In the future we may add a configuration setting that disables this
synchronous search indexing subscriber so that users of h.api (such as
our own h Python code) can subscribe to the AnnotationEvent and call the
search client method to index the annotations themselves - for example
from an asynchronous background process.