Trigger a <annotation> scope digest if the collapsible state of an annotation body changes
Excerpts update their overflow state outside of the Angular $digest
cycle. The usual way to deal with this is to wrap the update in a
$scope.$apply() call. However, this triggers a full $digest cycle, which
we want to avoid if many excerpts update their state at the same time -
eg. if we have a large number of annotation cards.

This commit instead explicitly invokes the $digest _only_ in the
affected annotation card and then only if the collapsible state of an
annotation body actually changed.