Update <excerpt> overflow state when excerpt is shown or hidden via ng-show
In addition to the content changing, the window resizing and media loading,
another way that the excerpt's overflow state can change is if it or a parent
element's visibility changes.

Since there isn't a generic, performant way to watch the visibility of DOM
elements, the change here adds a scope watch which tests for the presence of
the 'ng-hide' class used by the ng-show and ng-hide directives on parent
elements, since that is the only method of changing visibility that we actually
need to support at the moment.