Add the skeleton of a functional test suite
This commit adds the basic components of a functional test suite, which
tests the entire WSGI application as one unit.

These tests currently do not run on Travis. It's likely we won't want
them to run on Travis long-term, as it's likely the functional test
suite will end up rather slow. For now, they can be run manually with

    tox -e functional