Move Feature model into h.features.models
Move the Feature model into its own module. The set of current valid
flags is part of the model, so moves into this module.

I've made a couple of small changes that allow the details of the
FEATURES and FEATURES_PENDING_REMOVAL dictionaries to be hidden behind
the interface exposed by the model class, including a substantial
simplification to the Client class that results from reusing the
existing Feature.all() method rather than reimplementing its
functionality within the client.

The remove_old_flags function also moves into the model as a classmethod
so that everything that needs to know about the
FEATURES{,_PENDING_REMOVAL} dictionaries is in one place.

Added tests for `Feature.all()`, `Feature.description`