Move features client into h.features.client module
This moves the features client into its own module and substantially
simplifies the testing of this component.

The logic of returning only valid current features is already
implemented in the model class, so the Client doesn't need to duplicate
this functionality. As a result, the interaction between the client and
the model is reduced to a single "fetcher" callable, which is injected
for ease of testing.