Cache features data even when there are no features
This commit fixes a bug with the features client where removing all the
current feature flags from the application would result in the data from
the features table not being cached.

In practice, it's unlikely this would have caused any problems, as with
no features declared `client.enabled(...)` should not be called
anywhere, and `client.all()` is only called once per request.
Nevertheless, it's a clear bug in the implementation and is at least
confusing.