Add Postgres UpdateAnnotationSchema class
Add a new h.api.schemas.UpdateAnnotationSchema class for use when
validating an annotation update request for storage in postgres.
Based on AnnotationSchema as CreateAnnotationSchema is, and following
the same pattern.

Move the existing Elasticsearch UpdateAnnotationSchema class to
LegacyUpdateAnnotationSchema.

Change the update() view to use LegacyUpdateAnnotationSchema, the new
UpdateAnnotationSchema isn't used yet.