Fix reply notifications when 'postgres' is on
reply_template was crashing all over the place when passed a Postgres
annotation because it treats the annotation as a dict which does not
work with Postgres annotations.

The fix is mostly to use annotation class properties instead of
dictionary key access as this will work with both Elasticsearch and
Postgres annotations.

In one place logic had to be fixed: Postgres annotations don't have the
permissions access control list, but the relevant logic can be
implemented using the simpler shared and groupid properties instead.