Use `sudo docker` in docs
`docker` on its own will not work unless the user has taken additional
steps (which we don't document or suggest) to make docker work without
sudo. The error message won't be helpful either.

If the user has configured docker to run without sudo then they'll know
they've done that and can remove the sudo themselves (or not, it won't
do any harm).

This is consistent with all the other sudo commands in these docs.