Log the user out if they don't exist (#3231)
Problem:

It's possible for request.authenticated_userid to be set because the user has
a valid browser session but request.authenticated_user to be None at the same
time because the user doesn't exist in the database.

This happens when we delete a user account because we don't (and can't
easily) invalidate the user's session at delete time.

This leaves the application in an invalid state (contradiction between
request.authenticated_userid and request.authenticated_user, and
request.authenticated_userid referring to a user that no longer exists) and
parts of the code that don't expect this will crash.

Solution:

Ideally we would invalidate the user's sessions when we delete them but
this is difficult to do. So instead:

If a piece of code tries to access request.authenticated_user and the
user isn't found in the db then invalidate the user's session and
redirect the browser.

The user will be logged out.

We redirect to the same URL that was requested so that for example if you were
loading the group read page you'll get the logged-out version of the group read
page and not some other page.

We decided to do this when request.authenticated_user is requested because it
avoids doing an additional database lookup (to check that
request.authenticated_user would match what request.authenticated_userid says)
for requests that would not have needed request.authenticated_user.

This avoids an additional db lookup on every request. On the other hand
deleted users can still remain authenticated and do things as long as they
don't make a request that tries to use request.authenticated_user.

Add tests for the new solution and also clean up a few things in the
existing authenticated_user() tests:

- Introduce @authenticated_user_fixtures, apply the same fixtures to all
  authenticated_user() tests without repetition
- Make get_user a fixture to work with @authenticated_user_fixtures
- Use DummyRequest and authn_policy, more realistic test mocks
- Add test for the return value, was untested
- Small improvements to test docstrings

Fixes #3227