Return strings from document.title
Return either a single string or None from the Document.title property.

The Document.title property was returning a list of strings (the value
of each DocumentMeta is a list of strings). This seems wrong, and also
some of the tests seem written to expect strings and the legacy Document
objects return strings. So change the new Postgres Document.title to
return just a string as well.