Add auth.has_permission()
Remove duplicate and fragile authorization logic by adding an
auth.has_permission() function and using it.

This is for when you want to know if a user (other than the user making
the current request) has a certain permission on a certain object
(e.g. "can they read this annotation?"(