Return all canonical uris when expanding a canonical uri
Before we always only returned the passed-in uri when it ended up being
a canonical uri for the stored document. This works well since all
annotations will have the canonical uri as the `target[source]`.

The only issue is that when a website changes its canonical uri, then we
can't surface the annotations that were made previously because they
won't have the same canonical uri.

This commit fixes that behavior in returning all canonical uris when we
detect that the passed-in uri is a canonical uri.