Add conversation thread construction module
This is the first part of the new threading UI implementation
whose design is outlined at https://github.com/hypothesis/h/pull/3176/

build-thread implements the logic for constructing an immutable view
model for a list of conversations (a 'Thread') from a list of
annotations and replies plus an object describing the current UI state.

This view model is then rendered by an <annotation-thread> component.

Once the other parts of the new implementation are ready, this will
replace 'threading', 'thread-filter' and the JWZ script.