Move email generation out of `h.notification` package
This commit attempts to simplify the `h.notification` package, in
preparation for further overhaul, by moving all responsibility for email
generation out of the package.

Rather than the `generate_notifications` function yielding rendered
email data, we now have a business-logic-only function which checks to
see if a given annotation action merits a notification, and if so
yields a `Notification` object, which contains the data necessary for
rendering the notification, namely:

- the reply annotation,
- the user who made the reply annotation,
- the annotation being replied to,
- the user who made the annotation being replied to,
- the document object for the annotated page

The responsibility for actually generating the email has moved into a
new `h.emails` package, where `h.emails.reply_notification.generate` now
accepts the notification object as a parameters and returns the data
needed to construct an email.

As a result of the simplification of `h.notification.reply_template` in
this commit, the following modules have also been removed:

- `h.notification.gateway`
- `h.notification.notifier`
- `h.notification.types`