Don't require the "permissions" field
The general design of the API is that when updating an annotation if you
don't send a given field in the update request that field's value is
simply left unchanged. Any field can be omitted from API requests in
this way.

"permissions" was the one field that was required.

Allow "permissions" to be omitted in the same way as the other fields.

Also when creating an annotation, fields can be omitted and they will get
default values. Make "permissions" behave the same.