Update annotation and document relationships to allow eager loading
When loading annotations we now support eager loading of the documents
with its document uris and meta records. This should be done with
`sqlalchemy.orm.subqueryload` instead of `sqlalchemy.orm.joinedload` to
minimize the amount of data being transferred over the wire.

To load annotations with eager loading just the document:

    Annotation.query.options(subqueryload(Annotation.document)). \
        order_by(Annotation.updated.asc())

To load annotations with eager loading the document's document uris:

    Annotation.query.options(subqueryload(Annotation.document). \
                             subqueryload(Document.document_uris)). \
        order_by(Annotation.updated.asc())

To load annotations with eager loading the document's meta records:

    Annotation.query.options(subqueryload(Annotation.document). \
                             subqueryload(Document.meta)). \
        order_by(Annotation.updated.asc())

To load annotation with eager loading all the document's uris and meta
records:

    Annotation.query.options(
        subqueryload(Annotation.document).subqueryload(Document.document_uris),
        subqueryload(Annotation.document).subqueryload(Document.meta)
    ).order_by(Annotation.updated.asc())