Don't reset fields when updating an annotation
To be consistent with the behaviour of the legacy Elasticsearch
API/schemas/storage code, when an API client updates an annotation with
the 'postgres' flag on they can choose to send only some fields (e.g.
only text: "Hi") and only the values of those fields will be updated,
other fields will be left unchanged rather than being deleted from the
annotation or reset to default values.