Add "links" to protected fields
Add "links" to the list of protected fields (fields that are always
ignored if the client sends them in annotation create or update API
requests).

A "links" property was recently added to the annotations that we return
to the API: https://github.com/hypothesis/h/pull/3081

This is not a property that can be set or updated by the user. It's not
stored in the database. The links are _generated_ at annotation read
time.

Unfortunately this meant that, when the 'postgres' feature flag was on:

- The client would receive the "links" field in the annotation from the
  API

- The client would send back the unmodified "links" field in annotation
  update requests (it sends back all fields that it got from the API)

- UpdateAnnotationSchema, because its code doesn't recognise the "links"
  field, would then end up adding this field to the annotation "extra"
  field in the DB

We don't want to store these generated, unmodifiable links in the DB.

Adding "links" to the list of ignored fields fixes the problem.