Virtualize the annotation thread list
Make the sidebar handle "large" (ie. 50+) numbers of annotation threads
much better by virtualizing the thread list. Virtualizing means that we
only instantiate <annotation-thread> components for threads which are
either in or 'near' (above/below) the viewport. For the remaining
offscreen threads space is reserved by a couple of <li> elements above
and below the visible threads.