Interpret permissions more conservatively
Instead of considering an annotation to be private (not shared) if and
only if permissions['read'] == [userid], consider an annotation to be
_shared_ if and only if permissions['read'] == [groupid].

This shouldn't make any difference to the behaviour given the annotation
data that our own client sends when creating or updating annotations.

But when other potential clients call the API, this means that if the
given read permissions don't match what we expect we'll interpret the
annotation as private rather than as shared by default.