Fix SyntaxError exception when injecting sidebar
Fix an exception when injecting the sidebar caused by not setting the
`type` property on the <script> tag that contains the sidebar's JSON
config data.

This resulted in the browser trying to execute the JSON data as a JS
script. Injection still worked because the rest of the code still ran
and was able to read the JSON data from the script tag via the
textContent accessor.