Simplify and split up CLI package
In preparation for adding some new commands to make local development
more straightforward, this commit splits the `h.cli` package up so each
subcommand is defined in its own module. It also makes the following
breaking changes:

- The `annotool` command is removed. It was broken anyway, and we're
  very soon going to have an entirely different way of making updates to
  the annotation data, through normal alembic data migrations.

- The `token` command is removed. This was primarily useful before we
  had normal user API tokens.

- The `version` command is removed: use `hypothesis --version` instead.

- All other commands no longer accept a config file path as an argument.
  Instead, the `hypothesis` command has a top-level `--dev` flag which
  switches the behaviour of subcommands into a mode suitable for local
  development.

This set of changes is implemented using the `click` third-party
library, which is (to my knowledge) the easiest way to build
command-line applications which support nested subcommands.