Deduplicate tests
Now that we have TestCreateUpdateAnnotationSchema tests that are common
to both TestCreateAnnotationSchema and TestUpdateAnnotationSchema can go
in TestCreateUpdateAnnotationSchema, removing test code duplication.

I also found some tests that existed only in
TestCreateAnnotationSchema or only TestUpdateAnnotationSchema but that
applied to both CreateAnnotationSchema and UpdateAnnotationSchema and
moved those in to TestCreateUpdateAnnotationSchema as well.