Remove a bad test
It's now storage's job to merge extras posted in an annotation update
with extras already stored in the DB, schemas just passes any extras
that were posted along to storage. So "it overwrites extra fields in the extra
dict" (when the user posts a new value for an existing extra) no longer
makes sense as an UpdateAnnotationSchema test.