Fix EBADF crash with ws4py and gevent 1.1
We now override ws4py's run_application due to the websocket server
crashing with EBADF. A change in gevent (1.1) causes all sockets to be
closed when a WSGI handler returns. ws4py starts a new greenlet for
each new websocket connection used to return from the WSGI handler.
The fix is taken from [1] and waits for the greenlet to finish before
returning from the WSGI handler.

[1]: https://github.com/Lawouach/WebSocket-for-Python/pull/180

More information at:
- https://github.com/Lawouach/WebSocket-for-Python/issues/170
- https://github.com/gevent/gevent/issues/633