Add failing test case that reproduces the issue in #3278
Add a basic integration test for annotation anchoring and highlighting
with two cases, a simple case which works and another simple case which
reproduces the issue in #3278.

The test anchors annotations by triggering the 'annotationsLoaded' event
rather than calling Guest#anchor() in order to mirror what happens in
the actual app.