Anchor annotations sequentially
When anchoring annotations which overlap, inserting the highlight
for an annotation could fail if the DOM was mutated in-between the
annotation's location being resolved to a DOM Range and that DOM Range
being highlighted.

Given two annotations, #1 and #2 where #2's range is contained within #1
(eg. #1 annotates a sentence and #2 annotates a phrase within that
sentence), the following could happen:

 1. Guest#anchor() called with annotation #1
 2. Guest#anchor() called with annotation #2
 3. Annotation #1 resolved to range R1
 4. Annotation #2 resolved to range R2
 5. Range R1 highlighted. This wraps in a '<span class="annotator-hl">' tag
    and makes R2 invalid in the process.
 6. Range R2 highlighted. This then fails.

This commit resolves the issue by anchoring annotations sequentially, so
the sequence would now be:

 1. Guest#anchor() called with annotation #1
 2. Annotation #1 resolved to range R1
 3. Range R1 highlighted
 4. Guest #anchor() called with annotation #2
 5. Annotation #2 resolved to range R2
 6. Range R2 highlighted

It might be possible to optimize the process by only anchoring
annotations which overlap sequentially. This commit takes the simplest
approach and just anchors all annotations sequentially.

Fixes #3278