Make uri required when creating new annotations
The problem: we want to make a non-empty "uri" a required field when
creating new annotations, but _not_ when updating annotations.

I could easily do this by adding Python code to
CreateAnnotationSchema.validate(), but I would rather do things like
this that _can_ be done in JSONSchema in JSONSchema.

If I make "uri" required in the AnnotationSchema.schema class property,
because this is a class property and both CreateAnnotationSchema and
UpdateAnnotationSchema use an AnnotationSchema instance "uri" will now
be required both when creating and when updating annotations.

In general the JSONSchema validation that's done when creating and
updating cannot differ. Creating and updating validation can only differ
in the CreateAnnotationSchema.validate() and
UpdateAnnotationSchema.validate() methods in the code that runs after
AnnotationSchema.validate() is called.

The proposed solution:

1. Make AnnotationSchema.schema an instance attribute not a class
   attribute.

2. Make CreateAnnotationSchema add "uri" to "required" in the schema of
   its AnnotationSchema instance. This does not affect UpdateAnnotationSchema
   which has its own AnnotationSchema instance.

3. Because it uses the schema attribute, JSONSchema.validator now needs to be
   instantiated _after_ CreateAnnotationSchema has modified the schema.
   So move this from an instance attribute created in
   JSONSchema.__init__() (which happens before CreateAnnotationSchema
   modifies the schema) to an @property method which is not called until
   the validator is actually needed, when validate() is called.

   This means that an AnnotationSchema instance now creates a new
   validator each time its validate() is called, instead of creating
   just one and reusing it, but I think this is ok.

4. Make JSONSchema.schema an instance attribute as well, just to be
   consistent with AnnotationSchema.schema, there seems no point in
   mixing class and instance attributes, although I think we might get
   away with it.