Add BatchDeleter for cleaning up the ES index
It fetches all annotation ids currently indexed, then all annotation ids
in Postgres and compares them. The ones that are still in the index but
do not exist in Postgres (because they have been deleted and the
individual index delete celery task failed/didn't run) are then being
removed from the index.