Set document relationship on Annotation as viewonly
It wasn't possible to delete an annotation and then flush or commit the
changes to the database.
This is due to SQLAlchemy interpreting the relationship as a
many-to-many relationship. And even though the cascade option is
disabled, it tries to delete the many-to-many records. But our secondary
relationship is a join which does not have a delete method.

We don't even want SQLAlchemy to delete the "many-to-many" record, so
the correct fix is to set the viewonly option on this relationship.