Move python tests into tests/ directory
This commit moves all test code into the tests/ tree, and out of the
package directories.

This makes it easier to run tests against both library code (h.api, soon
to be memex) and application code (the rest of h), while following good
practices for both.

With some rearrangement of the repository (namely, moving the relevant
code into a src/ directory so it cannot appear on sys.path) we can
ensure that api/memex tests run against the *installed* version of the
package, which will serve as a check of the packaging configuration

For about the motivations for this pattern, see:

     https://hynek.me/articles/testing-packaging, and
     https://blog.ionelmc.ro/2014/05/25/python-packaging/