Send up to 200 replies per search request
Increase the upper limit on the number of reply annotations sent in a
single search request response with _separate_replies is True from 100
to 200.

This reduces the number of "Message not available" annotation cards seen
in the browser in place of reply annotations, on pages that have a lot
of replies.