Don't rely on whitespace insertion for correct spacing
The annotation action buttons are inline-block elements which means that
whitespace between the tags results in inserted whitespace in the
document when rendered. This rendered whitespace is roughly 4px wide.

Instead, switch the layout of the annotation actions div to use flexbox,
in which no such whitespace is inserted, and add 4px to the margin on
the buttons.