Make AnnotationHTMLPresenter work with Postgres annotations
And it no longer needs to support Elastic annotations.

The code now assumes that annotations have a .target_selectors property.

Since AnnotationHTMLPresenter now works with Postgres rather than
Elastic annotations, the field is now called target_uri rather than uri.

Since AnnotationHTMLPresenter now works with Postgres rather than
Elastic annotations it expects the annotation to be an object not a
dict.

We can assume that an annotation's properties have the right types now.

Postgres annotations have a target_selectors field instead of a target
field.

Although the client still sends them, fields other than "selector" in an
annotation's target_selectors are discarded and not stored.

Now that we've moved from the legacy Elasticsearch to the new Postgres
model an annotation's target_selectors property is a list of selectors
rather than a list of targets each containing a list of selectors as
before.