Use new TokenAuthenticationPolicy and remove cruft
This commit updates our primary AuthenticationPolicy so it delegates to
one or other of a SessionAuthenticationPolicy or the new
TokenAuthenticationPolicy depending on the request path. Both "upstream"
authentication policies use the same "groupfinder" callback to provide a
full list of effective principals for the authenticated user.

This allows to perform the following cleanups:

- h.auth.util.effective_principals turns into h.auth.util.groupfinder,
  and thus can be dramatically simplified: it no longer reimplements the
  default logic of Pyramid w.r.t. system principals.
- h.auth.util.group_principals is folded into groupfinder.
- h.auth.util.has_permission is no longer needed, with one minor change
  to the reply notification code.

This paves the way to further simplification and generalisation of the
AuthenticationPolicy, so that we can make changing authentication for
different routes in the application easily (in particular, I'm lining
this up for the websocket endpoint.)