Remove the target_links property from annotations
Now that we've moved from Elasticsearch to Postgres storage, when
creating or updating an annotation we:

a. Discard all targets that the client sends except the first
b. Discard all fields of the first target except 'selector'

In particular the 'source' field of the target, that used to be stored
in Elasticsearch, is no discarded and not stored in Postgres.

This makes the target_links property (which was a list of the 'source'
fields of all the annotations targets) no longer makes sense, so remove
it and remove the Atom feed code that was (pointlessly) still using it.