Remove the "links" field from the search index
Remove the "links" field of annotations from the search index.

This field did not exist in the search index back when we used the
legacy Elasticsearch storage, it only existed in the annotations
returned by the API, but it got accidentally added into the search index when
we moved to the new Postgres storage becayse the new search indexer uses
AnnotationJSONPresenter which is the presenter used for returning
annotations from the API.

Fundamentally we don't want the annotations returned by the API to have
to look exactly the same as the annotations in the search index - these
are two different things, and may evolve in different directions
according to different needs.

So add a new AnnotationSearchIndexPresenter and have the search indexer
use that instead of AnnotationJSONPresenter. Now the annotation formats
used by the API and search index are separated in the code and can be
modified independently.

Note that both AnnotationJSONPresenter and
AnnotationSearchIndexPresenter still use DocumentJSONPresenter for the
document sub-object of annotations - it has not been necessary to
separate out a DocumentSearchIndexPresenter yet.